chore: move computation logic for daily timeline updates to participant service
